dev-requirements.txt
alembic==1.8.1 ; python_version >= "3.8" and python_version < "4.0"
appnope==0.1.3 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "darwin"
asttokens==2.0.8 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.8" and python_version < "4.0"
backcall==0.2.0 ; python_version >= "3.8" and python_version < "4.0"
bandit==1.7.4 ; python_version >= "3.8" and python_version < "4.0"
beautifulsoup4==4.11.1 ; python_version >= "3.8" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.8" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.8" and python_version < "4.0"
coverage==6.5.0 ; python_version >= "3.8" and python_version < "4.0"
coverage[toml]==6.5.0 ; python_version >= "3.8" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.8" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.8" and python_version < "4.0"
dnspython==2.2.1 ; python_version >= "3.8" and python_version < "4.0"
docopt==0.6.2 ; python_version >= "3.8" and python_version < "4.0"
email-validator==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
executing==1.1.0 ; python_version >= "3.8" and python_version < "4.0"
filelock==3.8.0 ; python_version >= "3.8" and python_version < "4.0"
flake8==5.0.4 ; python_version >= "3.8" and python_version < "4.0"
flask-migrate==3.1.0 ; python_version >= "3.8" and python_version < "4.0"
flask-sqlalchemy==2.5.1 ; python_version >= "3.8" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
gitdb==4.0.9 ; python_version >= "3.8" and python_version < "4.0"
gitpython==3.1.27 ; python_version >= "3.8" and python_version < "4.0"
greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.8" and python_version < "4.0"
identify==2.5.5 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
importlib-metadata==4.13.0 ; python_version >= "3.8" and python_version < "3.10"
importlib-resources==5.9.0 ; python_version >= "3.8" and python_version < "3.9"
infinity==1.5 ; python_version >= "3.8" and python_version < "4.0"
iniconfig==1.1.1 ; python_version >= "3.8" and python_version < "4.0"
intervals==0.9.2 ; python_version >= "3.8" and python_version < "4.0"
invoke==1.7.3 ; python_version >= "3.8" and python_version < "4.0"
ipython==8.5.0 ; python_version >= "3.8" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
jedi==0.18.1 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
mako==1.2.3 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "4.0"
matplotlib-inline==0.1.6 ; python_version >= "3.8" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.8" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.8" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.8" and python_version < "4.0"
parso==0.8.3 ; python_version >= "3.8" and python_version < "4.0"
pbr==5.10.0 ; python_version >= "3.8" and python_version < "4.0"
pexpect==4.8.0 ; python_version >= "3.8" and python_version < "4.0" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.8" and python_version < "4.0"
platformdirs==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
pre-commit==2.20.0 ; python_version >= "3.8" and python_version < "4.0"
prompt-toolkit==3.0.31 ; python_version >= "3.8" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.8" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.8" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.8" and python_version < "4.0"
pycodestyle==2.9.1 ; python_version >= "3.8" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.8" and python_version < "4.0"
pyflakes==2.5.0 ; python_version >= "3.8" and python_version < "4.0"
pygments==2.13.0 ; python_version >= "3.8" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "4.0"
pytest-cov==3.0.0 ; python_version >= "3.8" and python_version < "4.0"
pytest-watch==4.2.0 ; python_version >= "3.8" and python_version < "4.0"
pytest==7.1.3 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.8" and python_version < "4.0"
python-testmark==1.0 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==65.4.1 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.8" and python_version < "4.0"
soupsieve==2.3.2.post1 ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy-utils==0.38.3 ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy==1.4.41 ; python_version >= "3.8" and python_version < "4.0"
stack-data==0.5.1 ; python_version >= "3.8" and python_version < "4.0"
stevedore==4.0.0 ; python_version >= "3.8" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.8" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.8" and python_version < "4.0"
traitlets==5.4.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.3.0 ; python_version >= "3.8" and python_version < "4.0"
validators==0.20.0 ; python_version >= "3.8" and python_version < "4.0"
virtualenv==20.16.5 ; python_version >= "3.8" and python_version < "4.0"
watchdog==2.1.9 ; python_version >= "3.8" and python_version < "4.0"
wcwidth==0.2.5 ; python_version >= "3.8" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.8" and python_version < "4.0"
wtforms-alchemy==0.18.0 ; python_version >= "3.8" and python_version < "4.0"
wtforms-components==0.10.5 ; python_version >= "3.8" and python_version < "4.0"
wtforms==3.0.1 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.8.1 ; python_version >= "3.8" and python_version < "3.10"
